+++ crates/matrix-sdk-crypto/src/identities/device.rs
@@ -481,6 +481,11 @@ impl Device {
 
         Ok(raw_encrypted)
     }
+
+    /// Whether or not the device is a dehydrated device.
+    pub fn is_dehydrated(&self) -> bool {
+        self.inner.inner.dehydrated.unwrap_or(false)
+    }
 }
 
 /// A read only view over all devices belonging to a user.